backend/eamiddleware/requirements.txt
asgiref
certifi
charset-normalizer
click
colorama
Django
djangorestframework
geometry
idna
py4j
pyspark
requests
six
sqlparse
tzdata
urllib3
kafka-python
redis
pandas
numpy
